start-stop-daemon: package the debian version of this tool
The busybox version does not support the --no-close option, without which our
apps, which do not deamonize, so need to be started with --background, end up
logging to /dev/null.

This recipe is based on the outdated recipe in meta-gumstix. The source file
was extracted from the dpkg 1.16.16 tarball and defines and includes adjusted
to allow standalone compilation.