python: disable over-assertive assert in python GC
Due to some mismatch between the Yocto versions of pygobject and gst-python,
specifically to do with floating references in GObject, we are hitting an
assert in the GC due to a missing reference. While this is indicative of a
problem elsewhere, the GC code below the assert handles this situation safely
so aborting the application is not appropriate, if nothing else, because it
makes it almost impossible to find out what the real cause of the problem is.
So we replace the assert with a warning message.