stdlicense.bbclass: extract logic for copying common license into a class
Some of our packages do not have sources per se, so no license file (this is
true of packages with genuinely no source, like wifi-setup, or packages that
that wrap a pre-compiled bits, such as thin). In cases where the correct
license is among the common Yocto licenses, we can just copy it from there
rather than having to add a copy to sources.