radiodan-apps: workaround wiringpi/wiring-pi brokenness
wiring-pi 'helpfully' pulls in and tries to build the wiringpi applications,
unfortunately the wiringpi makefiles are broken (overriding CC and other
variables with hardcoded values), so the binaries npm will build are for host,
not target. We don't really care about these, as we package wiringpi separately
where the broken makefiles can be patched, so just blow away the binary
artefacts.