radiodan-buttons: work around wiring-pi crosscompilation issues
wiring-pi does not crosscompile, the basic problem is that it does a git
checkout of wiringpi, which it then attempts to build, but wiringpi cannot
be crosscompiled without patching the Makefiles because it hard codes "CC=gcc"

There is no nice way to fix this as there is no way to hook up into the
npm install midway to patch. The workaround is to run npm install, then patch
the relevant makefiles and then rebuild with npm rebuild.